Update parser.c
I had to change  the env function as it was not working on linprog also modified the path search so it works with arguments. There is both error checking for environment variables and path searches.  Not sure why but the environment variable was working on local machine but not on linprog so I fixed it. Path search has been tested with echo and ls -l