Update parser.c
Background Processes work as well as jobs. Some bugs to fix are that the file redirection when outputting a file is only read only idk why. Also when using pipes for background processes the output will sometimes pop out on the command line. However this does not affect the next input